INS-934: rename heavy package to heavyserver
