+++ ledger/heavyserver/heavysync_test.go
@@ -14,18 +14,17 @@
  *    limitations under the License.
  */
 
-package heavy
+package heavyserver
 
 import (
 	"testing"
-
-	"github.com/insolar/insolar/testutils"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/insolar/insolar/core"
 	"github.com/insolar/insolar/instrumentation/inslogger"
 	"github.com/insolar/insolar/ledger/storage/storagetest"
+	"github.com/insolar/insolar/testutils"
 )
 
 func TestHeavy_SyncBasic(t *testing.T) {
